Move the primal-dual perturbation handler into pounce-common
Pure refactor, no behaviour change. Makes the IPOPT-ported
PdPerturbationHandler reachable from pounce-qp, which is the
prerequisite for #410.

Context: the NLP filter-IPM solves 105/138 Maros-Meszaros instances
where the active-set QP solves 46/138. Part of that gap is structural
(an IPM never commits to an active set, so LICQ of A_W never arises),
but part is simply that pounce-algorithm already implements the full
Wachter-Biegler (2006 §3.1) correction -- delta_x, delta_s, delta_c,
delta_d, with a degeneracy probe and a state machine -- and pounce-qp
implements none of it, shifting only the H block. Benzi, Golub &
Liesen (Acta Numerica 14, 2005, Thm 3.8) prove an H-block shift can
never repair a rank-deficient constraint block, which is what these
instances have.

The dependency runs pounce-algorithm -> pounce-qp, so the QP side
could not reach the handler. Rather than duplicate 674 lines of ported
logic, it moves down to pounce-common, which both already depend on.

The only coupling was diagnostics, now behind a `PerturbationSink`
trait with defaulted methods, so pounce-common stays dependency-light
(it deliberately avoids tracing; the caller supplies that). Two
methods carry the info-string and lg(rg) writes, two more the
env-gated POUNCE_DBG_PERT trace. pounce-algorithm keeps its module
path as a re-export plus an `IpoptDataSink` adapter over
IpoptDataHandle, so nothing else in that crate changed.

I tried a hand-rolled fixed delta_c in pounce-qp first. It made things
worse -- SearchDirectionBecomesTooSmall became
InfeasibleProblemDetected on QSCTAP1, QSCAGR7 and Q25FV47, a false
infeasibility claim on feasible problems, and Q25FV47 went 19.7s ->
75.7s. IPOPT's delta_c is not a constant: it is delta_cd_val * mu^exp,
applied only after a degeneracy probe identifies the Jacobian as the
culprit, with delta_d alongside. That is why this ports the state
machine rather than reinventing a piece of it.

pounce-common + pounce-algorithm + pounce-qp: 43 binaries, 673 tests,
0 failures.
