+++ crates/pounce-common/src/lib.rs
@@ -12,6 +12,7 @@ pub mod diagnostics;
 pub mod exception;
 pub mod journalist;
 pub mod options_list;
+pub mod pd_perturbation;
 pub mod reg_options;
 pub mod style;
 pub mod tagged;
